rename "accessories" to "decorators"
