+++ src/chart-decorators/point.js
